requirements.txt
click==7.1.1
mutagen==1.44.0
opencc-python-reimplemented==0.1.5